Builds a Workset of Hathi Trust vol IDs by querying the Workset Builder 2.0

Queries the SOLR endpoint of the Workset Builder 2.0 (beta) at https://solr2.htrc.illinois.edu/solr-ef/. This API is experimental, and so this function can fail at any time if the API changes.

workset_builder(
  token,
  genre,
  title,
  name,
  imprint,
  pub_date,
  lang = "en",
  volumes_only = TRUE,
  token_join = c("AND", "OR"),
  genre_join = c("AND", "OR")
)

Arguments

token

The tokens to search for in the HathiTrust Extracted Features files. Can be a vector of characters, e.g., c("liberal", "democracy"); if a vector, the results are interpreted using the value of token_join -- by default AND, so that the query will find all volumes where the tokens appear in the same page, but not necessarily next to each other (in that case, all pages containing both "liberal" and "democracy"). If token_join is "OR" then the query will find all volumes where either of the tokens appear. Search is case-insensitive; phrases can be included (e.g., "liberal democracy"), and the database will then return matches that are separated by hyphens (i.e., pages that contain the token "liberal-democracy").

genre

A genre string, e.g. "Fiction" or "Not Fiction" in the Hathi Trust full metadata. Can be a set of genre strings, e.g., c("dictionary","biography"), which are interpreted using the value of genre_join (by default "AND", so the metadata must contain both genre strings; if "OR", the query will look for volumes that contain any of the genre strings).

title

Title field. Multiple words will be joined with "AND"; can be a phrase (e.g., "Democracy in America").

name

Names associated with the book (e.g., author). Multiple terms will be joined with "AND"; can be a phrase (e.g., "Alexis de Tocqueville").

imprint

Imprint information (e.g., published). Multiple terms will be joined with "AND"; can be a phrase (e.g., "University of Chicago Press").

pub_date

Publication date in Hathi Trust metadata. Can be a range, e.g., 1800:1900, or a set of years, e.g., c(1800, 1805).

lang

Language. Default is "en" (english). Right now this function can only search one language at a time; if you wish to search for terms in more than one language, create multiple worksets and bind them together.

volumes_only

If TRUE (the default), returns only volume IDs plus a count of the number of times the tokens appear in the volume; FALSE returns both volume and page IDs where the tokens are found.

token_join

The logical connector for different tokens. Default is "AND"; the query will ask for all pages where all tokens occur. "OR" means the query will ask for all pages where any of the tokens occur.

genre_join

The logical connector for different genre strings. Default is "AND"; the query will ask for all volumes containing all the selected genre strings. "OR" means the query will ask for all volumes where any of the genre strings occur.

Value

A tibble with volume_ids, number of occurrences of the terms in the volume, and if volumes_only is FALSE a column for page ids.

Examples

if (FALSE) {
res <- workset_builder("democracy", genre = c("biography", "dictionary"))
# Better to run two separate queries so the genres can be distinguished per
# volume, but one can in principle use "OR" to find volumes that contain
# "democracy" in either genre.
res <- workset_builder("democracy", genre = c("biography", "dictionary"),
       genre_join = "OR")

# All volumes that mention "tylenol" and "paracetamol" in the same page
res <- workset_builder(c("tylenol", "paracetamol"), volumes_only = FALSE)

# All volumes mentioning "demagogue" published between 1800 and 1900
res <- workset_builder("demagogue", pub_date = 1800:1900)

# All volumes mentioning "demagogue" with "Tocqueville" and "Reeve"
# in the author field
res <- workset_builder("demagogue", name = c("Tocqueville", "Reeve"))

# All volumes with "Tocqueville" in the author field
res <- workset_builder(name = "Tocqueville")
}