Reads the metadata of a single downloaded Hathi Trust extracted features file

Note that if you want to extract the metadata of more than one Hathi Trust ID at a time, it is best to simply query the Workset Builder database using get_workset_meta. Using get_workset_meta also guarantees you'll get a rectangular data with HT ids in the rows. It is also possible to get simple metadata for large numbers of htids by downloading the big hathifile using download_hathifile and then filtering it.

get_hathi_meta(htid, dir = getOption("hathiTools.ef.dir"))

Arguments

htid

The Hathi Trust id of the item whose extracted features files are to be downloaded.

dir

The directory where the file is saved. Defaults to getOption("hathiTools.ef.dir"), which is just "./hathi-ef/" on load. If the file does not exist, this function will first attempt to download it.

Value

a tibble with metadata.

Examples

if (FALSE) {
# Download the 1862 version of "Democracy in America" by Tocqueville and get
# its metadata

tmp <- tempdir()

get_hathi_meta("mdp.39015001796443", dir = tmp)
}