Get metadata for a set of Hathi Trust IDs

Queries the SOLR endpoint of the Workset Builder 2.0 (beta) at https://solr2.htrc.illinois.edu/solr-ef/ to download volume metadata. This API is experimental, and so this function can fail at any time if the API changes.

get_workset_meta(
  workset,
  metadata_dir = getOption("hathiTools.metadata.dir"),
  cache = TRUE
)

Arguments

workset

A workset of htids, generated by workset_builder from Hathi Trust's SOLR endpoint. One can also pass a data frame with a column labeled "htid" and containing valid Hathi Trust htids, or a character vector of htids (though the function will complain with a warning).

metadata_dir

The directory used to cache the metadata file. Defaults to getOption("hathiTools.metadata.dir"), which is just "./metadata" on loading the package.

cache

Whether to cache the resulting metadata as a CSV. Default is TRUE. The name of the resulting metadata file is generated by appending an MD5 hash (via digest::digest) to the string "metadata-", so each metadata download will have a different name.

Value

A tibble::tibble with the Hathi Trust metadata for all the volumes in the workset or the vector of htids.

Details

Also, be mindful that dowloading a large number of metadata records can take quite some time. Even a few hundred records can take a few minutes.

Examples

if (FALSE) {
workset <- workset_builder(name = "Tocqueville")
meta <- get_workset_meta(workset)

## We can also pass a vector of htids:
meta <- get_workset_meta(workset$htid)
}